refactor: split binding implementations into .cpp files
