+++ src/scripting/lua_engine.hpp
@@ -2,7 +2,6 @@
 #include <memory>
 #include <vector>
 
-#define SOL_ALL_SAFETIES_ON 1
 #include <sol/sol.hpp>
 
 #include "binding_module.hpp"
