+++ src/scripting/script_event_bridge.hpp
@@ -6,7 +6,6 @@
 #include <unordered_map>
 #include <vector>
 
-#define SOL_ALL_SAFETIES_ON 1
 #include <sol/sol.hpp>
 
 #include "lua_engine.hpp"
