+++ src/scripting/script_scheduler.hpp
@@ -3,7 +3,6 @@
 #include <memory>
 #include <vector>
 
-#define SOL_ALL_SAFETIES_ON 1
 #include <sol/sol.hpp>
 
 #include "lua_engine.hpp"
